requirements.txt
asgiref==3.2.3
confusable-homoglyphs==3.2.0
Django==3.0.3
django-markdownx==3.0.1
django-registration==3.1
django-taggit==1.2.0
Markdown==3.1.1
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
